python_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
twilio==8.10.0
python-telegram-bot==20.7
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
plotly==5.18.0
python-dateutil==2.8.2
pytz==2023.3.post1
aiohttp==3.9.1
asyncio==3.4.3 